Move Security cog to backend extension
